Modified request files

+++ app/Http/Requests/StoreCategoryRequest.php
@@ -22,10 +22,10 @@ class StoreCategoryRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'created_by' => ['nullable', 'numeric', 'min:1'],
+            'created_by' => ['nullable', 'integer', 'min:1'],
             'name' => ['required', 'string'],
             'slug' => ['nullable', 'string'],
-            'active' => ['nullable', 'numeric'],
+            'active' => ['nullable', 'integer'],
         ];
     }
 }
